Window: the device reset lives in Options under INTERFACE, with one shared confirm dialog
