+++ src/OpenXLR.UI/MainWindow.axaml.cs
@@ -131,42 +131,8 @@ public partial class MainWindow : Window
         ProfileNameBox.Text = "";
     }
 
-    /// <summary>Small in-app confirmation dialog; true when the user accepts.</summary>
-    private async Task<bool> ConfirmAsync(string title, string message, string yesLabel = "Overwrite")
-    {
-        var yes = new Button { Content = yesLabel, Background = Avalonia.Media.Brush.Parse("#a03434") };
-        var no = new Button { Content = "Cancel" };
-        var dialog = new Window
-        {
-            Title = title,
-            SizeToContent = SizeToContent.WidthAndHeight,
-            WindowStartupLocation = WindowStartupLocation.CenterOwner,
-            CanResize = false,
-            Background = Avalonia.Media.Brush.Parse("#1d2027"),
-            Content = new StackPanel
-            {
-                Margin = new Avalonia.Thickness(18),
-                Spacing = 14,
-                Children =
-                {
-                    new TextBlock { Text = message, TextWrapping = Avalonia.Media.TextWrapping.Wrap, MaxWidth = 380 },
-                    new StackPanel
-                    {
-                        Orientation = Avalonia.Layout.Orientation.Horizontal,
-                        Spacing = 8,
-                        HorizontalAlignment = Avalonia.Layout.HorizontalAlignment.Right,
-                        Children = { no, yes },
-                    },
-                },
-            },
-        };
-        var done = new TaskCompletionSource<bool>();
-        yes.Click += (_, _) => { done.TrySetResult(true); dialog.Close(); };
-        no.Click += (_, _) => { done.TrySetResult(false); dialog.Close(); };
-        dialog.Closed += (_, _) => done.TrySetResult(false);
-        await dialog.ShowDialog(this);
-        return await done.Task;
-    }
+    private Task<bool> ConfirmAsync(string title, string message, string yesLabel = "Overwrite")
+        => Dialogs.ConfirmAsync(this, title, message, yesLabel);
 
     private void OnProfileLoad(object? sender, RoutedEventArgs e)
     {
@@ -176,13 +142,6 @@ public partial class MainWindow : Window
     private void OnProfileDelete(object? sender, RoutedEventArgs e)
     {
         if ((sender as Button)?.Tag is string name) _vm.DeleteProfile(name);
-    }
-
-    private async void OnResetDevice(object? sender, RoutedEventArgs e)
-    {
-        if (!await ConfirmAsync("Reset device to defaults?", _vm.ResetDescription, yesLabel: "Reset"))
-            return;
-        _vm.ResetDevice();
     }
 
     private void OnPickDevice(object? sender, RoutedEventArgs e)
